100% Code Coverage for GroupsController
